feat: Создаёт тестовые данные
