+++ src/const.ts
@@ -1,4 +1,4 @@
-export type City = 'Paris'|'Cologne'|'Brussels'|'Amsterdam'|'Hamburg'|'Dusseldorf';
+import { City } from './types/city';
 
 export const Setting = {
   OffersCount: 312
@@ -14,6 +14,7 @@ export const CITIES: City[] = [
 ];
 
 export const DEFAULT_CITY: City = CITIES[0];
+
 export enum AppRoute {
   Main = '/',
   Login = '/login',
