+++ src/mocks/offers.ts
@@ -1,0 +1,44 @@
+import { Offer } from '../types/offer';
+
+export const offers: Offer[] = [
+  {
+    id: 1,
+    img: 'img/apartment-01.jpg',
+    priceValue: 100,
+    rating: 80,
+    placeCardName: 'Beautiful flat in the center of the city',
+    placeCardType: 'Flat'
+  },
+  {
+    id: 2,
+    img: 'img/apartment-01.jpg',
+    priceValue: 220,
+    rating: 100,
+    placeCardName: 'Beautiful & luxurious apartment at great location',
+    placeCardType: 'Apartment'
+  },
+  {
+    id: 3,
+    img: 'img/apartment-01.jpg',
+    priceValue: 20,
+    rating: 60,
+    placeCardName: 'Cozy hostel room',
+    placeCardType: 'Hostel'
+  },
+  {
+    id: 4,
+    img: 'img/apartment-01.jpg',
+    priceValue: 300,
+    rating: 90,
+    placeCardName: 'Luxurious apartment in the city center',
+    placeCardType: 'Apartment'
+  },
+  // {
+  //   id: 5,
+  //   img: 'img/apartment-01.jpg',
+  //   priceValue: 120,
+  //   rating: 30,
+  //   placeCardName: 'Beautiful & luxurious apartment at great location',
+  //   placeCardType: 'Apartment'
+  // }
+];
