--- src/mocks/place-cards.ts
@@ -1,42 +1,0 @@
-export const placeCards = [
-  {
-    id: 1,
-    img: 'img/apartment-01.jpg',
-    priceValue: 120,
-    rating: 80,
-    placeCardName: 'Beautiful & luxurious apartment at great location',
-    placeCardType: 'Apartment'
-  },
-  {
-    id: 2,
-    img: 'img/apartment-01.jpg',
-    priceValue: 120,
-    rating: 100,
-    placeCardName: 'Beautiful & luxurious apartment at great location',
-    placeCardType: 'Apartment'
-  },
-  {
-    id: 3,
-    img: 'img/apartment-01.jpg',
-    priceValue: 120,
-    rating: 60,
-    placeCardName: 'Beautiful & luxurious apartment at great location',
-    placeCardType: 'Apartment'
-  },
-  {
-    id: 4,
-    img: 'img/apartment-01.jpg',
-    priceValue: 120,
-    rating: 90,
-    placeCardName: 'Beautiful & luxurious apartment at great location',
-    placeCardType: 'Apartment'
-  },
-  {
-    id: 5,
-    img: 'img/apartment-01.jpg',
-    priceValue: 120,
-    rating: 30,
-    placeCardName: 'Beautiful & luxurious apartment at great location',
-    placeCardType: 'Apartment'
-  }
-];
