+++ src/types/offer.ts
@@ -1,0 +1,8 @@
+export type Offer = {
+  id: number;
+  img: string;
+  priceValue: number;
+  rating: number;
+  placeCardName: string;
+  placeCardType: string;
+};
