+++ src/types/place-card-props.ts
@@ -1,0 +1,7 @@
+export type PlaceCardProps = {
+  img: string;
+  priceValue: number;
+  rating: number;
+  placeCardName: string;
+  placeCardType: string;
+}
